Move allowRoundingHacks to Internals interface https://bugs.webkit.org/show_bug.cgi?id=87328
Reviewed by Hajime Morita.

.:

* Source/autotools/symbols.filter: Add allowRoundingHacks symbol filter.

Source/WebCore:

Add allowRoundingHacks function, because it is able to work in the
cross-port way by means of the Internals interface.

In addition, a function is added in order to restore default setting values.

No new tests, since we are improving here the infra-structure for testing
a specific method.

* testing/Internals.cpp:
(WebCore::Internals::reset):
(WebCore):
(WebCore::Internals::resetDefaultsToConsistentValues):
(WebCore::Internals::allowRoundingHacks):
* testing/Internals.h:
(Internals):
* testing/Internals.idl:

Source/WebKit2:

* win/WebKit2.def: Add setAllowsRoundingHacks() to symbol filter.

Tools:

Remove allowRoundingHacks functions, because it is able to work in the
cross-port way through the Internals interface.

* DumpRenderTree/LayoutTestController.cpp:
(LayoutTestController::staticFunctions):
* DumpRenderTree/LayoutTestController.h:
(LayoutTestController):
* DumpRenderTree/blackberry/LayoutTestControllerBlackBerry.cpp:
* DumpRenderTree/efl/LayoutTestControllerEfl.cpp:
* DumpRenderTree/gtk/LayoutTestControllerGtk.cpp:
* DumpRenderTree/mac/LayoutTestControllerMac.mm:
* DumpRenderTree/win/LayoutTestControllerWin.cpp:
* DumpRenderTree/wx/LayoutTestControllerWx.cpp:

LayoutTests:

Invoke allowRoundingHacks function from internals, because it is able to work in the
cross-port way via the Internals interface.

* platform/mac/fast/text/rounding-hacks-expansion.html:
* platform/mac/fast/text/rounding-hacks.html:


Identifier: 105402@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118657 268f45cc-cd09-0410-ab3c-d52691b4dbfc