Factor DeferrableOneShotTimer out of GraphicsContextCG/GeneratorGeneratedImage https://bugs.webkit.org/show_bug.cgi?id=87824
Reviewed by Simon Fraser.

DeferrableOneShotTimer is a timer which takes a class and method to call (as Timer does),
and a delay. It exposes restart(), stop(), and isActive(), and none of the rest of
the TimerBase interface. On restart(), the timer is started with the given delay,
unless it's already running, in which case a flag is set to automatically restart
the timer when it fires instead of calling the given method. This provides a
low-overhead way to implement a deferrable timer.

No new tests, refactoring.

* platform/Timer.h:
(DeferrableOneShotTimer):
(WebCore::DeferrableOneShotTimer::DeferrableOneShotTimer): New class.
(WebCore::DeferrableOneShotTimer::restart):
(WebCore::DeferrableOneShotTimer::fired):
* platform/graphics/GeneratorGeneratedImage.cpp:
(WebCore::GeneratorGeneratedImage::invalidateCacheTimerFired):
* platform/graphics/GeneratorGeneratedImage.h:
(WebCore::GeneratorGeneratedImage::GeneratorGeneratedImage): Make use of DeferrableOneShotTimer.
* platform/graphics/cg/GraphicsContextCG.cpp:
(WebCore::SubimageCacheWithTimer::SubimageCacheWithTimer): Make use of DeferrableOneShotTimer.
(WebCore::SubimageCacheWithTimer::invalidateCacheTimerFired):


Identifier: 105666@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@118942 268f45cc-cd09-0410-ab3c-d52691b4dbfc