ComposedShadowTreeWalker should support traversing nodes in an orphaned shadow subtree. https://bugs.webkit.org/show_bug.cgi?id=87493
Reviewed by Dimitri Glazkov.

Source/WebCore:

ComposedShadowTreeWalker assumed that a visited shadow root is
always assigned to a shadow insertion point since it only
traverses nodes which are rendered.  But there is an exceptional
use case such as an event dispatching. Some events, such as a
'click' event, may happen in an orphaned shadow subtree.  In such
cases, traversal might start with a node in an orphaned shadow
subtree.  So ComposedShadowTreeWalker can not assume that visited
shadow root is always assigned to a shadow insertion point.

This patch only fixes ComposedShadowTreeWalker.
ComposedShadowTreeParentWalker will be fixed in another patch with
an event dispatching test.

Test: fast/dom/shadow/composed-shadow-tree-walker.html

* dom/ComposedShadowTreeWalker.cpp:
(WebCore::ComposedShadowTreeWalker::traverseParentBackToYoungerShadowRootOrHost):

LayoutTests:

* fast/dom/shadow/composed-shadow-tree-walker.html:


Identifier: 105767@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@119069 268f45cc-cd09-0410-ab3c-d52691b4dbfc