ASSERT(m_fontSelector->document()) is bogus and should be removed https://bugs.webkit.org/show_bug.cgi?id=88053
Reviewed by Abhishek Arya.

This ASSERT was introduced in http://trac.webkit.org/changeset/97402
together with a branch that handled the case of the ASSERT firing.
This ASSERT fires when running tests on Android (which runs with
ASSERTs enabled). The ASSERT appears to be bogus, so this patch removes
it.

* css/CSSSegmentedFontFace.cpp:
(WebCore::CSSSegmentedFontFace::getFontData):


Identifier: 105886@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@119207 268f45cc-cd09-0410-ab3c-d52691b4dbfc