[chromium] Remove redundant setNeedsCommit when prepareToDraw fails https://bugs.webkit.org/show_bug.cgi?id=88246
Reviewed by James Robinson.

Source/WebCore:

When prepareToDraw fails it explicitly calls setNeedsCommit, but the
scheduler already takes care of this for it. When a frame is not
successfully drawn, the scheduler state machine sets the needsCommit
bit itself.

This behaviour is covered by CCSchedulerStateMachineTest.TestFailedDrawSetsNeedsCommitAndDoesNotDrawAgain

* platform/graphics/chromium/cc/CCLayerTreeHostImpl.cpp:
(WebCore::CCLayerTreeHostImpl::prepareToDraw):

Source/WebKit/chromium:

* tests/CCLayerTreeHostImplTest.cpp:


Identifier: 106113@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@119447 268f45cc-cd09-0410-ab3c-d52691b4dbfc