Assigning to a static property should not change iteration order https://bugs.webkit.org/show_bug.cgi?id=88401
Reviewed by Geoff Garen.

A specific iteration order is not defined by the spec, but test-262 somewhat tenuously
requires that it is at least stable, e.g. ch10/10.4/10.4.2/S10.4.2_A1.1_T1.js

Whilst it is not clear that this behavior really arises from the specification, it
would seem like common sense to conform to this.

The problem here is that we allow properties in the structure to shadow those in the
static table, and we iterate the properties in the structure first - which means that
as values of existing properties are modified, their iteration order changes too.

The easy fix is to iterate the properties from the static table first. This has a
further benefit, since it will mean that user added properties will come after those
present in the static table (respected the expected insertion-order).

Source/JavaScriptCore:

* runtime/JSObject.cpp:
(JSC::JSObject::getOwnPropertyNames):
    - Iterate static properties first.

LayoutTests:

* fast/js/property-iteration-expected.txt: Added.
* fast/js/property-iteration.html: Added.
* fast/js/script-tests/property-iteration.js: Added.
(arrayEqual):
    - Added test case.



Identifier: 106276@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@119623 268f45cc-cd09-0410-ab3c-d52691b4dbfc